pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "eagleeye"
version = "0.1.0"
description = "EagleEye - Real-time scam intelligence CLI tool - Your personal scam radar"
authors = ["EagleEye Team <contact@eagleeye.dev>"]
readme = "README.md"
homepage = "https://eagleeye.dev"
repository = "https://github.com/eagleeye/eagleeye"
documentation = "https://eagleeye.dev"
keywords = ["security", "scam", "fraud", "intelligence", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Monitoring",
]

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
httpx = "^0.25.2"
pydantic = "^2.5.0"
pydantic-settings = "^2.0.3"
pyyaml = "^6.0.1"
click = "^8.1.7"
asyncio-mqtt = {version = "^0.16.1", optional = true}
aiofiles = "^23.2.1"
spacy = "^3.7.0"
textblob = "^0.17.1"
scikit-learn = "^1.3.0"
numpy = "^1.24.0"
pandas = "^2.0.0"
nltk = "^3.8.0"
phonenumbers = "^8.13.0"
requests = "^2.31.0"
urllib3 = "^2.0.0"
python-dateutil = "^2.8.0"
validators = "^0.22.0"
tldextract = "^5.0.0"
beautifulsoup4 = "^4.12.2"
joblib = "^1.3.2"
speechrecognition = "^3.10.0"
pydub = "^0.25.1"
librosa = "^0.10.1"
geoip2 = "^4.7.0"

[tool.poetry.extras]
notifications = ["asyncio-mqtt"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
coverage = "^7.3.2"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
eagleeye = "eagleeye.cli:app"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]
